Ignore duplicate path/field/token links.
This is occasionally important when dealing with arrays in metadata,
where each element is processed separately and not subject to token
deduplication; as a result, we end up attempting to insert duplicate
path/field/token links. By ignoring hese, we ensure that the index
builds, and we also save space by trimming out duplicate tokens.